+++ lib/core/flame/components/entities/mobs/rock_golem.dart
@@ -4,7 +4,6 @@ import 'package:defend_your_flame/core/flame/components/entities/disappear_on_de
 import 'package:defend_your_flame/core/flame/components/entities/draggable_entity.dart';
 import 'package:defend_your_flame/core/flame/components/entities/configs/entity_config.dart';
 import 'package:defend_your_flame/core/flame/components/entities/enums/idle_time.dart';
-import 'package:defend_your_flame/core/flame/components/entities/mixins/has_draggable_collisions.dart';
 import 'package:defend_your_flame/core/flame/components/entities/mixins/has_idle_time.dart';
 import 'package:defend_your_flame/core/flame/helpers/entity_helper.dart';
 import 'package:defend_your_flame/helpers/misc_helper.dart';
@@ -12,19 +11,19 @@ import 'package:flame/collisions.dart';
 import 'package:flame/components.dart';
 import 'package:flutter/rendering.dart';
 
-class RockGolem extends DraggableEntity with DisappearOnDeath, HasIdleTime, HasDraggableCollisions {
+class RockGolem extends DraggableEntity with DisappearOnDeath, HasIdleTime {
   static final EntityConfig _rockGolemConfig = EntityConfig(
     entityResourceName: 'rock_golem',
     defaultSize: Vector2(90, 64),
     defaultScale: 1.4,
-    walkingConfig: AnimationConfig(frames: 9, stepTime: 0.15),
-    attackingConfig: AnimationConfig(frames: 11, stepTime: 0.11),
+    walkingConfig: AnimationConfig(frames: 10, stepTime: 0.15),
+    attackingConfig: AnimationConfig(frames: 11, stepTime: 0.13),
     dragConfig: AnimationConfig(frames: 8, stepTime: 0.16),
-    idleConfig: AnimationConfig(frames: 8, stepTime: 0.16),
+    idleConfig: AnimationConfig(frames: 8, stepTime: 0.15),
     dyingConfig: AnimationConfig(frames: 13, stepTime: 0.14),
     damageOnAttack: 15,
     goldOnKill: 8,
-    baseWalkingSpeed: 24,
+    baseWalkingSpeed: 22,
     totalHealth: DamageConstants.fallDamage * 4,
     dragResistance: 0.52,
     timeSpendIdle: TimeSpendIdle.minimal,
@@ -32,7 +31,7 @@ class RockGolem extends DraggableEntity with DisappearOnDeath, HasIdleTime, HasD
   );
 
   late final RectangleHitbox _hitbox =
-      EntityHelper.createRectangleHitbox(size: Vector2(34, 33), position: Vector2(32, 48), anchor: Anchor.bottomCenter);
+      EntityHelper.createRectangleHitbox(size: Vector2(35, 33), position: Vector2(44, 64), anchor: Anchor.bottomCenter);
 
   RockGolem({super.scaleModifier, super.modifiedWalkingSpeed}) : super(entityConfig: _rockGolemConfig);