+++ lib/core/flame/components/entities/mobs/skeleton.dart
@@ -13,7 +13,7 @@ class Skeleton extends DraggableEntity with DisappearOnDeath {
     defaultSize: Vector2(22, 33),
     attackingSize: Vector2(43, 37),
     defaultScale: 1.65,
-    walkingConfig: AnimationConfig(frames: 13, stepTime: 0.09),
+    walkingConfig: AnimationConfig(frames: 13, stepTime: 0.092),
     attackingConfig: AnimationConfig(frames: 18, stepTime: 0.1),
     dragConfig: AnimationConfig(frames: 4, stepTime: 0.2),
     dyingConfig: AnimationConfig(frames: 15, stepTime: 0.07),